controller: trigger_vbss_move: fix radio uid
The current way the Radio UID is read has a pernicious bug. After
being used to store the BSSID, the 'value' variable is re-used to
store the Radio UID. At the same time, the amxd_object_get_parent()
call used to retrieve the radio object succeeds, but only gets the
parent *BSS* object (as the log line indicates the Radio instance is
the grandparent of the BSS instance, not the parent). Because the
return value of the next amxd_object_get_param() call is not checked,
the flow continues with 'value' containing the BSSID (its previous
value) instead of the Radio UID.

Since we already have the Radio UID and BSSID in the database, fix it
by using the database instead of the datamodel.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>