bpl: platform_db: set the band type of wlan interfaces
Set the band type of the WLAN interfaces in
prplmesh platform db configuration file.
The name "band" is OpenWRT common option name, taken from
the following link:
https://openwrt.org/docs/guide-user/network/wifi/basic
Basically, there are two option names that determines
the band type: "band" and "hwmode", but "hwmode" is
deprecated, as stated in the link.

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>