controller: subscribe to CLIENT_SECURITY_CONTEXT_RESPONSE_MESSAGE
The controller needs to subscribe to the Client Security Context
Response message if it wants to handle it. Handlers for this message
were added in commit 530a90d9adea4a08183a2fa6cfc4eb65d96c5733 but the
subscription to the message was not added, so add it now.

While at it, add a debug message when we receive it so that when
manually testing we know it's going to be handled.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>