controller: vbss_task: set the active move state when creating it
The comment above the line that adds to the active moves mentions
setting the state to CLIENT_SEC_CTX but it's actually never set, so it
defaults to the INIT state.

Add a new constructor to sMoveEvent so that we can actually set the
sate when adding to the active moves.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>