common: Change BrokerClient factory to return shared_ptr
In a follow-up commit, the BrokerClient will be shared with the traffic
separation.
Technically we can leave it unique_ptr, and just pass the Agent context,
to use it through it. But this approach was meant to be temporary only,
and be refactored at the end of the Unified Agent epic.
It is generally a better approach to pass what is needed to another
component than give a context used as a global variable.

Since the Autoconfiguration task is the module that will pass the
BrokerClient to the TrafficSeparation class in follow-up commit, and it
is anyway using the Agent context, change the BrokerClient to public.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>