Agent: Change TrafficSeparation class to a non-static class
Until now the traffic separation class was used as a static class.
Change the class to a non-static class, since now it will need to hold
the BrokerClient which is a common resource.
Also, from now it will only be used only in the autoconfiguration task
so there is no point in it will be static.

While at it, change the "apply_traffic_separation" function name to
"apply_policy".

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>