Agent: Remove bSTA/bBSS interfaces from bridge monitoring
When traffic separation is configured, prplMesh creates a VLAN interface
assigned to the configured primary VLAN, linked for each bSTA and bBSS
it creates.
In general, the transport does not return a message to the interface it
was received, but since there is a VLAN interface linked to the
original interface, the transport sends it back to the VLAN interface,
and this ends with the packet going back to the link it was received on.

To prevent this, when defining a VLAN interface linked to another
interface, need to exclude the original interface from the transport
monitoring, so we will not be able to send and receive messages to/from
that interface.

Remove bSTA and bBSS from transport monitoring when linking them a VLAN
interface.
Removed old code that was supposed to do that and was commented out.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>