topology_task: override MultiApProfile according to certification program
In Topology Response, profile in MultiApProfile TLV is set
to MULTIAP_PROFILE_3 by default. For R2 certification,
profile is expected to be MULTIAP_PROFILE_2 in Topology
Response Message.

This behaviour is seen while running MAP-4.13.4 R2 tests
with Mediatek as CTT-Agent2.

If certification_mode is enabled, override the multiAP
profile with respect to certification program that will be
configured through ucc command.

Signed-off-by: Nandhini Murugesan <nmurugesan@maxlinear.com>