controller: db: set default values of channel_select_task
Add fix to set the default value of the parameters
channel_select_task and dynamic_channe_select_task to
true so that they run as expected. These tasks on running
correctly will allow the controller database to update
itself with right values such that it can be seen in the
output of bml_conn_map.

PPM-2393

Signed-off-by: badhri_s <bsrinivasan@maxlinear.com>