controller: save the non-parsed association frame per station
Up to now we stored the parsed association frame. While it can be
useful for cases where we only want to use the content of the
association frame, it makes it difficult to re-use it to send the
complete frame on the network as the AssocReqFrame class doesn't have
a way to serialize it. We could add a mechanism to serialize
it, but the (re-)association frames are complex enough for it to be
time-consuming (and error-prone) to write.

To make it easier to send it, store the raw association frame in a
vector. For places that need to use the content of the association
frame, it can still easily be parsed beforehand.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>