agent: handle VBSS move preparation request and response
When an agent receives a VBSS move preparation request, it should send
a move preparation response containing the same client info TLV, to
acknowledge the message.

If the agent has any agreement with the station (e.g. block acks that
have been negotiated), it should also terminate them.

For now, we only send the move preparation response and leave
terminating the agreements for later. This way, the flow can continue
on the controller side in the meantime.

Closes PPM-2404

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>