topology task: controller: add VS DeviceInformation TLV
Since the Non-VS Device Information TLV does not have the distinction
between a VAP and an actual AP (radio), we need a VS TLV to pass this
information on to the BMl task (via the BML event).

If the connected controller is a prplMesh controller we need to send
this VS-TLV in order to inform the BML task.
For non-prplMesh controller this isn't an issue since the BML is part
of the controller-only tasks.

Resolves PPM-1977.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>