agent: ap_manager: register ext_events handlers when adding a BSS
When a new BSS is added, we need to register ext_events handlers for
it.

Use the newly-introduced register_ext_events_handler() method for
that.

To make the code flow easier to follow, introduce add_bss() that both
asks the ap_wlan_hal to create the BSS and register its ext_events
handlers.

Closes PPM-2406

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>