controller: add a note about saving the tunneled association frame
There is currently nothing that prevents the tunneled association
frame to reach the controller before the topology notification for the
same station.

Because of that, if we try to save the association frame in the DB it
will fail because the station does not exist yet.

Since we don't have an immediate solution for that add a comment so
that at least we know about it.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>