controller: send client capability queries to prplMesh agents
Currently, the prplMesh controller has a different behavior regarding
client capabilities when talking to a prplMesh agent, than it does
when talking to a non-prplMesh agent. For prplMesh agents, the client
capabilities are retrieved from a vendor specific TLV in the topology
notification, and no client capability query is sent.

This is problematic for R1 prplMesh agents, as it means we won't get
an association frame for the station to save in the controller (the
association frame is contained in the client capability report, but
not in the vendor specific TLV of the topology response).

To keep the same behavior between prplMesh and non-prplMesh agents,
don't add the client capabilities in the datamodel yet in
'verify_sta_association()'. Instead, send a client capability query
like we do for non-prplMesh agents. Once we receive the client
capability report, the capabilities will be added to the
datamodel. This way, we can also save the association frames that
comes in the client capability report for clients connected to
prplMesh agents.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>