agent: ap_manager: append radio index to the VBSS interface name
We currently cannot move a VBSS from one radio to another one on the
same agent, because the interface name we use would be the same (since
it's based only on the BSSID).
This is an unnecessary limitation that we want to overcome.

To make the interface name more unique, we need to include something
specific to the radio in it. Unfortunately, Linux interface names are
limited to 15 characters, so with 12 characters already used by the
BSSID we are quite limited in choice.

Ideally, we would use some unique number of each AP Manager (e.g. the
position in the list of radio managers in the agent). This however
requires too many changes to be reasonable to do right now. In order
to still be able to make progress, for now use the radio index from
the interface name, but leave a comment to fix it in a new task.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>