agent: ap_manager: fix unreachable code when registering ext_events
The condition "ext_event_fd < 0" is within the 'ext_event_fd > 0', so
the portion of code it wraps is unreachable.

Fix it by checking first for 'ext_event_fd < 0', and continue if
needed (this removes the need for an else clause which is closer to
our coding style).

Note that the case where the file descriptor would be 0 is now handled
in a different way than before. There is however no reason to treat 0
in a different way than 1.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>