ci: test-gw-repeater: use the docker-test tag
With only the 'docker-build' tag, the job is picked up by the existing
'Alderaan-docker-privileged' runner (in addition to the other ones)
which doesn't have docker available. As a consequence, when running on
that runner the test fails as can be seen in [1].

In addition, it is also picked-up by the
'devalore-prplmesh-builder-shell' runner, on which creating new Docker
networks is no longer possible (see [2]).

To prevent that, introduce a new 'docker-test' tag meant for (shell)
runners capable of running tests (other than Boardfarm tests) using
docker containers created by the job itself. The new tag has already
been added to the relevant runners in the Gitlab CI settings of the
prplMesh project.

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/3430658409
[2]: https://prplfoundationcloud.atlassian.net/browse/PPM-2408

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>