framework: tlvf: add ieee_80211 sMacHeader
This new file describes the (beginning of the) 802.11 MAC header as
well as enums to be used with it.

This is a preparatory commit for sending custom action frames, for
which we will need to specify the MAC header.

Note that the MAC header is not in a TLV format, but we use the TLV
framework for it because it already provides a mechanism to define and
use bitfields.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>