tools: docker: clang-format.sh: fix running from a git worktree
tools/docker/clang-format.sh has the annoying limitation that it
cannot be run from a git worktree, as it uses git to list files and
the worktree directory alone doesn't have a "full" git folder but only
a gitlink to the common directory.

To overcome it, check if git has a common folder different from the
local git folder. If it does, mount it as a volume as well.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>