controller: agent: dm: nbapi: PPM-2409, implement QoS configuration from custom DM trigger
Add NBAPI trigger for the QoS configuration from controller to agents based on a
new custom configuration in DM. The new node Device.WiFi.DataElements.Configuration.QoS
is the configuration to be shaped and sent to the agents.The agents provide their feedback
according to the standard TR-181 DM tree. Thus, configuration and feedback are clearly
distinguished.
NOTE: The new QoS configuration reuse an existing Configuration node. Idealy this should
be move aside of the official tree, under X_PRPL_ORG.Configuration

Signed-off-by: Francois Gerin <francois.gerin@mind.be>