bwl: backhaul_manager: advance if back haul interface is already connected
bugfix for the case where STA is connected but backhaul manager is still waiting for WPS for the connection to start.

This sends a reassociate command to the STA which is already connected triggering back haul manager to advance its state from WAIT_WPS to CONNECTED (then OPERATIONAL).

There is another work https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/commit/29b18bc8f9874962b0f1cf27850f3de51e969981 to solve this issue.
That might be more handy though but was not working out of the box. Adapting that on to current head was not trivial.
The current solution (re-associating) was chosen instead because it was an easy way to make sure the state machine advances to the right state.

PPM-1717

Signed-off-by: Oguz Ozhan <oguz.ozhan@mind.be>