bwl: nl80211_client: set the station authorized when adding it
When adding the station, it won't be able to communicate with the rest
of the network until we authorize it.

Since in our case (i.e. for the VBSS feature) we want the station to
move to the target AP without any disruption for the user, add the
authorized flag when we add the station.

Note that this should also be removed once we add the station through
hostapd (PPM-2358).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>