boardfarm_tests: channel_selection: adjust num of radios
The number of radios in the agent entity depends on the device
in which the boardfarm test is running (prplmesh_compose.py,
prplmesh_cgr.py, prplmesh_prplwrt,py, etc...)
In all the classes that derives from ALEntity, two Radio objects
are created. However, in the previous commits, a new Radio object
was added to the ALEntityDocker, which runs inside a docker, and
it is also is used in the pipeline tests.
Since the third radio is created only on the dockerized device,
we can't access it in the ChannelSelection test when it runs on
devices that have only two radios.
Hence, refactor the ChannelSelection test so it won't assume
a fixed number of radios, but rather iterate over the radios list.

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>