boardfarm_tests: nbapi_radio_bh_sta: fix mismatch of ruids
The Backhaul STA Capability Report message contains one
or more Backhaul STA Radio Capabilities TLVs. Inside each of
these TLVs there is a RUID value, that needs to be match with one
of the agent's radio mac. However, these RUIDs and the
agent radio macs are not always arranged in the same order.
Thus, this commit solves the case when the RUID and the
agent's mac address are not arranged in the same order.

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>