transport: fix raw socket creation failure for interfaces in bridge
If transport fails to create a raw socket for a network
interface in bridge, add it to the pending network interfaces
list.

If UP event is received for an interface, not part of the
interfaces monitored by transport and if it is part of the
pending network interfaces list, create a raw socket for the
interface, transport will start monitoring it.

PPM-2424

Signed-off-by: Nandhini Murugesan <nmurugesan@maxlinear.com>