framework: uci: Match radio ifname between prplmesh config and wireless.
Add the parameter exclude_hostap_iface to the prplmesh_db.
Add a fix such that it compares the value of 'hostap_iface' in the
prplmesh config with the 'ifname' value in the wireless config, if
found match interpret that as radio and do not add its ssid and other
related information in the controller's M2.
The above is done only if the parameter exclude_hostap_iface is set
to true.

PPM-2427

Signed-off-by: bsrinivasan <bsrinivasan@maxlinear.com>