common: bwl: add sta_acceptlist_modify()
This new method allow to add to/remove from/clear an accept list.

Note that for the nl80211 implementation, sta_allow()/sta_deny() already does
only removing from the deny list and adding to the deny list
respectively. With that in mind, it would make sense to take the
opportunity to rename sta_allow() to sta_denylist_remove() and
sta_deny() to sta_denylist_add().
However, the WHM and dwpal implementation have a different
behavior. The WHM implementation can add to/remove from both the deny
list and accept list, depending on whether the station should be
allowed or denied.
The dwpal implementation has its own STA_ALLOW and DENY_MAC custom
hostapd commands, which are different from the open source hostapd
deny lists/accept lists.

To avoid adding to the confusion (because the different BWLs don't
have the same behavior for sta_allow() and sta_deny()), sta_allow()
and sta_deny() are not renamed.

Actual implementations for the new method will be added in follow-up
commits.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>