controller: vbss_task: increment TXPN
During a VBSS move, there's a small temporal window during which the
TXPN can change after we've already gotten it from the BWL layer.

If that occurs, the STA being moved will drop packets from the
destination BSS as it will be seen a potential replay attack (new BSS
has a TXPN less than reality).

To prevent this, we increase the TXPN so that it's arbitrarily larger
than what we originally knew it to be.

The current amount to increase it by, 1000, was derived emperically, and
depends on the speed at which the whole VBSS movement workflow can
happen, so it is subject to change in the future.

Signed-off-by: Tucker Polomik <t.polomik@cablelabs.com>