agent: ap_manager: add station to allow list when creating a VBSS
When creating a VBSS, add the station to an allow list. As a
consequence, broadcast probe request from any other station will not
be answered, to try to discourage them from trying to connect.

For stations that still attempt to connect by sending an
authentication request, the request will be rejected.

Denying the connection for other stations is even more important when
unicast beacons are used (which will be done in the next commit).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>