agent: ap_manager: set unicast beacons when creating a VBSS
We only configure unicast beacons if the BSS creation is the result of
a move. If it's the very first creation of a VBSS, we want to use
broadcast beacons first (as we don't necessarily know the final MAC
the station will have when it associates), and switch to unicast once
the station is connected (which will be done in PPM-2431).

Note that after setting beacon_da we have to call UPDATE_BEACON (which
is done in this commit) so that hostapd takes it into account.

This change depends on the following merge request in prplOS:
https://gitlab.com/prpl-foundation/prplos/prplos/-/merge_requests/281

Closes PPM-2192

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>