agent: vbss: do not deauthenticate unknown stations during VBSS creation
When we create a VBSS, there is a (very short) time window where the
station could send us data frames when the BSS has already been
created, but the station has not been added yet.

The default behavior in that case is to deauthenticate the station,
which we want to avoid in our case. Unfortunately, there is no way to
eliminate this race condition by adding the station while the BSS
doesn't exist yet, as the kernel doesn't make it possible to use
NL80211_CMD_NEW_STATION before NL80211_CMD_START_AP is called.

Set the (hostapd) option to not deauthenticate unknown stations when
we create the VBSS. Once the station has been added, we can restore
the default behavior of deauthenticating unknown stations.

Closes PPM-2418

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>