media_type : Refactor get_802_11_media_type function
Currently, the Wi-Fi standard is guessed from band and max bandwidth, which is
not always accurate. In addition, the implementation contains some errors as for
instance 160MHz is not a valid bandwidth in 2.4GHz band.

Using the HT/VHT/HE support flags already available in DB could ensure a more reliable
retrieval of the Wi-Fi standard, and thus of the media type.

Closes PPM-2657

Signed-off-by: Sylvain LE ROUX <prplmesh.opensources@sagemcom.com>