agent: add more logging to the topology task
From the logs, it's currently not clear why the topology response
would not be sent (i.e. which TLV fails to be added).

Add error logging, so that from the logs we can better know what is
actually done.

While at it, also add debug logging for the associated clients TLV, so
that we can know at a glance if the TLV is going to be added, and for
which client.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>