common: bwl: return ext event fd in add_bss and add_interface
When adding a new BSS, the ap_manager, which is responsible of
registering external event handlers for the new BSS, has no way to
know the new interface's external events file descriptor.

Up to now we used the last file descriptor of the container. However
that approach is wrong, as invalidated file descriptors can be re-used
anywhere inside the fds container.

To fix it, change the add_bss and add_interface APIs to return the
external event file descriptor for the newly-created interface.
In the AP manager, use it to register the external events handlers.

Note that in a follow-up commit, the file descriptor will also be
saved by the AP manager to be able to remove handlers on BSS removal.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>