bwl: nl80211: add_station: emit STA_Connected
hostapd does not emit a AP-STA-CONNECTED event if a station is manually
added via NL80211_CMD_NEW_STATION.

If manually adding a station succeeds, generate a STA_Connected event in
the BWL which will percolate up to the ap_manager.

The ap_manager will then enqueue the message, causing the
son_slave_thread to send a topology notification message to the
controller.

This allows the controller to update it's data model with the new
station that was manually added.

Signed-off-by: Tucker Polomik <t.polomik@cablelabs.com>