controller: take AL-MAC into account when adding/removing BSS
When the VBSS feature is used, 2 BSSes with the same BSSID can
co-exist for a short period of time, on different agent.

This is problematic for the controller, as get_bss() currently only
takes the BSSID into account to find a BSS in the DB. As a
consequence, when a VBSS is moved the controller does not create a new
BSS on the target agent when it reports it, as it thinks the BSS
already exists (since it finds the BSS with the same BSSID on the
source agent).

To fix it, add an optional 'al_mac' argument to get_bss() so that we
can search for a BSS on a specific agent. The new argument is
optional, as although in multiple cases we do have access to the
AL-MAC (since it's often the source address of a message), there are
still cases where we currently don't have access to it.

add_vap(), update_vap() and add_node_station() are similarly updated
so that when we want to add a new BSS, we first search for it on the
specific agent for which we want to create it instead of search on all
agents.

Fixes PPM-2447

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>