agent: remove external event handlers on BSS destruction
Now that we get the external events file descriptor at the BSS
creation, save it so we can use it later.

When removing a BSS, also remove the external event handlers.

This fixes the following errors that occur when a VBSS is created for
the second time on the same agent:
WARNING 02:35:59:294 <3069302112> beerocks_event_loop_impl.cpp[69] --> Requested to add FD (31) to the poll, but it's already there
ERROR 02:35:59:294 <3069302112> ap_manager.cpp[2912] --> Unable to register handlers for external event fd 31
ERROR 02:35:59:295 <3069302112> ap_manager.cpp[2872] --> Failed to register ext_events handlers for the new BSS!
ERROR 02:35:59:295 <3069302112> ap_manager.cpp[742] --> Failed to add a new BSS!

Fixes PPM-2448

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>