beerocks_message: add bridge name in wifi credentials update request
Introduce bridge_name in
ACTION_APMANAGER_WIFI_CREDENTIALS_UPDATE_REQUEST message
that is sent from autoconfiguration task to the  AP Manager
to configure the SSID credentials (sent in WSC M2 message)
in the platform.

This is to ensure correct bridge is configured in the
hostapd.conf file so that interfaces will be added to the right
bridge.

PPM-2475

Signed-off-by: Nandhini Murugesan <nmurugesan@maxlinear.com>