agent: task: link metric collection task changes
It is observed that we are trying to create ap metric query for
preparing ap metric response late when periodicity is set. When
compared elasped time in sec, which does not give better acuracy.
For example when reporting interval is 2 sec, we are ignoring se-
-nding  query to monitor when time  elasped is 1 sec but when we
check the same in ms it is around 1999 ms. Which does not look ok
wrt MAP-R4  strick time  constraint. So added grace period in ms
(150 ms) to send the query even time elasped in sec is 1 sec but
in ms it is in between 1850 to 1999.

PPM-2474

Signed-off-by: Ravi Kumar <rnandusao@maxlinear.com>