controller: db: fix update_unassociated_station_stats
Previously, we would first check if a given unassociated station was
already in the data model via a MAC address Ambiorix look-up.

If the unassociated station was found in the data model already, it's
stats would be updated.

However, if the unassociated station was being added for the first time,
it's new data model path would be ignored and only it's index would be
appended to an empty string object, which we would then try to do
Ambiorix transactions on.

This fixes the data model path resolution so this function can now
handle updating a previously known unassociation station, and adding one
for the first time followed immediately by updating it in the DM.

While we're in this code, refactor to prefer early-return on error
conditions to reduce cyclomatic complexity.

Signed-off-by: Tucker Polomik <t.polomik@cablelabs.com>