agent: ap_manager: send the VBSS response before the topology update
From the controller's point a view, a critical code path for the VBSS
move to be efficient is the time from which it requests the creation
of a VBSS on the target AP, to the time it request its deletion from
the source AP.

When analyzing logs on the controller side, we realized that a lot of
time is spent by the controller handling topology messages and the
'sACTION_CONTROL_HOSTAP_VAPS_LIST_UPDATE_NOTIFICATION' message that
'handle_aps_update_list()' sends, before it handles the VBSS creation
response.

To make the overall move faster, move the sending of the creation
response to before sending the topology/interface updates. This way,
the controller is going to handle the VBSS response first, allowing it
to proceed to sending the destruction request sooner.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>