agent: ap_manager: give AP some time before resetting deauth behavior
It was noticed during manual testing that when there is a high amount
of traffic between the AP and the station, the kernel still reports
unexpected frames from the station for some time after the station has
been marked associated and authorized. Supposedly, this could be
because after we mark the station as associated, the kernel still has
a queue of frames that it was handling as from an unknown station.

To prevent this from deauthenticating our VBSS station, give the AP
some more time before resetting no_deauth_unknown_sta.
We determined empirically that the kernel typically reports the
spurious frames within the same second. To be on the safe side, set
the default grace period to 2 seconds.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>