controller: nbapi: TriggerVBSSDestruction: find radio correctly
To traverse upwards in a data model object such as "Foo.1.Bar.2." from
an object pointing at "Bar", amxd_object_get_parent must be called twice
to get to "Foo".

Only calling it once and then querying the "ID" field would just yield
the BSSID ID again, which we would erroneously send as the radio UID.

Now, we send the BSSID to destroy as well as the RUID the BSS belongs
to.

Signed-off-by: Tucker Polomik <t.polomik@cablelabs.com>