controller: vbss_task: increase the TX PN even more when moving a VBSS
After some more experimentation, it turns out when there is a high
amount of traffic (tested with an 11n station), increasing by 1000 is
definitely not enough, and 10000 is close to not being enough.

To be on the safe side in this scenario, increase it by 100x what it
was before.

It has to be noted that with the same test setup, increasing the TX PN
by 1000000 (i.e. 1000x what it was before this change) also worked.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>