agent: ap_manager: configure the accept list if the station was associated
The accept list was removed because it prevents the station from
connecting when a VBSS is created for the first time (for any station,
in case we cannot know the station MAC in advance).

Add it back, but only when the station was previously associated.

This way, after the first move, other stations will be ignored if they
try to connect.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>