bwl: nl80211_client: set aggr_mode to manual when sending a delba
During our tests, we realized it's sometimes possible for the source
AP to re-negociate block acks after we sent the delba, but before we
had the time to destroy the source BSS. When this happens, the move
takes a very long time to complete, as both sides are still expecting
specific block acks in reply.

To make sure this does not happen, we would ideally configure the
source BSS to disable aggregation negotiation through the kernel (see
PPM-2191). For this to be possible, we need a kernel >=5.10, which we
will only get in prplos-next.

For now, use the debugfs entry that is available (and enabled by
default in prplos) for ath10k instead.

Note that this should all go away as part of PPM-2191.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>