nl80211_client_impl: add prplMesh <-> nl80211 translations
The shape of the client capability data that the kernel
expects via nla_put* is different from what's parsed/stored inside
prplMesh.

Add ieee80211_{v}ht_capabilities structs that are of the correct layout
for the kernel to accept, and add functions to translatae from the
association frame data that prplMesh has to the kernel-approved data.

Signed-off-by: Tucker Polomik <t.polomik@cablelabs.com>