monitor: nl80211: close station-sniffer socket on error
When we encounter an error (e.g. because station-sniffer crashed,
restarted, etc), prplMesh currently stops getting unassociated RSSIs
completely, because the socket is setup once and not reset in such a
case.

To fix it, reset the socket in case we encounter an error.

Additionally, we also want to close the file descriptor when we're
done with it. Ideally we would create an RAII class for it, but that
is more involved (it requires more changes). As a compromise, a make
the file descriptor a unique_ptr with a custom deleter, so that we
don't have to worry about closing it ourselves.

While at it, add some debug logs around the socket handling, so that
we can tell from the logs what happens when debugging.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>