agent: link_metrics_collection_task: call erase_client redundantly
When a VBSS move occurs, the sACTION_APMANAGER_CLIENT_ASSOCIATED_NOTIFICATION
sent during nl80211_client::add_station should take care of removing this
station from the source AP's DB. There's a call in son_slave_thread
to remove the newly associated client MAC from all radios, which it does.
But when this codepath is taken, somehow, the client MAC is back in the
source radios associated_clients list. Likely due to some sort of race condition, somewhere.
Since we're short on time for the VBSS demo, just blindly erase the station
MAC from the radio's associated_client list so we can make unassociated
link metrics requests regardless of which BSS the station is associated
to.

Signed-off-by: Tucker Polomik <t.polomik@cablelabs.com>