giant blob commit:
Fixes a ton of unassociated station data model problems.

The 'sawtooth' problem in the UI is resolved here.

Fixed a race condition between a station connected event and a VBSS move
success event in VBSS manager.

This commit needs to be broken up quite a lot, but this is a 'save
state' to be demo-ready-ish.

Signed-off-by: Tucker Polomik <t.polomik@cablelabs.com>